+++ platform/android/MapLibreAndroidTestApp/src/androidTest/java/org/maplibre/android/testapp/offline/OfflineManagerTest.kt
@@ -3,6 +3,8 @@ package org.maplibre.android.testapp.offline
 import android.content.Context
 import androidx.test.internal.runner.junit4.AndroidJUnit4ClassRunner
 import androidx.test.rule.ActivityTestRule
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.runTest
 import java.io.IOException
 import java.util.concurrent.CountDownLatch
 import org.junit.After
@@ -50,6 +52,7 @@ class OfflineManagerTest : AppCenter() {
         latch.await()
     }
 
+    @OptIn(ExperimentalCoroutinesApi::class)
     @Test(timeout = 30_000)
     fun combinedTest() {
 
@@ -59,18 +62,19 @@ class OfflineManagerTest : AppCenter() {
 
         val latch1 = CountDownLatch(1)
         rule.activity.runOnUiThread {
-            FileUtils.CopyFileFromAssetsTask(
-                rule.activity,
-                object : FileUtils.OnFileCopiedFromAssetsListener {
-                    override fun onFileCopiedFromAssets() {
-                        latch1.countDown()
-                    }
-
-                    override fun onError() {
-                        throw IOException("Unable to copy DB file.")
-                    }
-                }
-            ).execute(TEST_DB_FILE_NAME, FileSource.getResourcesCachePath(rule.activity))
+            var copied = false
+            runTest {
+                copied = FileUtils.copyFileFromAssetsTask(
+                    rule.activity,
+                    TEST_DB_FILE_NAME,
+                    FileSource.getResourcesCachePath(rule.activity)
+                )
+            }
+            if (copied) {
+                latch1.countDown()
+            } else {
+                throw IOException("Unable to copy DB file.")
+            }
         }
         latch1.await()