+++ platform/android/MapLibreAndroidTestApp/src/main/java/org/maplibre/android/testapp/activity/FeatureOverviewActivity.kt
@@ -5,17 +5,17 @@ import android.content.Intent
 import android.content.pm.PackageInfo
 import android.content.pm.PackageManager
 import android.content.res.Resources.NotFoundException
-import android.os.AsyncTask
-import android.os.Build
 import android.os.Bundle
-import android.util.Log
 import android.view.View
-import androidx.annotation.RequiresApi
 import androidx.annotation.StringRes
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.SimpleOnItemTouchListener
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import org.maplibre.android.testapp.R
 import org.maplibre.android.testapp.adapter.FeatureAdapter
 import org.maplibre.android.testapp.adapter.FeatureSectionAdapter
@@ -37,7 +37,6 @@ class FeatureOverviewActivity : AppCompatActivity() {
     private var sectionAdapter: FeatureSectionAdapter? = null
     private var features: List<Feature>? = null
 
-    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_feature_overview)
@@ -65,22 +64,26 @@ class FeatureOverviewActivity : AppCompatActivity() {
     }
 
     private fun loadFeatures() {
-        try {
-            LoadFeatureTask().execute(
-                packageManager.getPackageInfo(
-                    packageName,
-                    PackageManager.GET_ACTIVITIES or PackageManager.GET_META_DATA
+        lifecycleScope.launch(Dispatchers.IO) {
+            try {
+                features = loadFeaturesTask(
+                    packageManager.getPackageInfo(
+                        packageName,
+                        PackageManager.GET_ACTIVITIES or PackageManager.GET_META_DATA
+                    )
                 )
-            )
-        } catch (exception: PackageManager.NameNotFoundException) {
-            Timber.e(exception, "Could not resolve package info")
+                withContext(Dispatchers.Main) {
+                    onFeaturesLoaded(features)
+                }
+            } catch (exception: PackageManager.NameNotFoundException) {
+                Timber.e(exception, "Could not resolve package info")
+            }
         }
     }
 
-    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     private fun onFeaturesLoaded(featuresList: List<Feature>?) {
         features = featuresList
-        if (featuresList == null || featuresList.isEmpty()) {
+        if (featuresList.isNullOrEmpty()) {
             return
         }
         val sections: MutableList<FeatureSectionAdapter.Section> = ArrayList()
@@ -100,7 +103,7 @@ class FeatureOverviewActivity : AppCompatActivity() {
             FeatureAdapter(features!!)
         )
         sectionAdapter!!.setSections(sections.toTypedArray())
-        recyclerView!!.adapter = sectionAdapter
+        recyclerView.adapter = sectionAdapter
     }
 
     private fun startFeature(feature: Feature) {
@@ -114,59 +117,52 @@ class FeatureOverviewActivity : AppCompatActivity() {
         outState.putParcelableArrayList(KEY_STATE_FEATURES, features as ArrayList<Feature>?)
     }
 
-    private inner class LoadFeatureTask : AsyncTask<PackageInfo?, Void?, List<Feature>>() {
-        protected override fun doInBackground(vararg p0: PackageInfo?): List<Feature>? {
-            val features: MutableList<Feature> = ArrayList()
-            val app = p0[0]
-            val packageName = FeatureOverviewActivity.javaClass.`package`.name
-            val metaDataKey = getString(R.string.category)
-            if (app != null) {
-                for (info in app.activities) {
-                    if (info.labelRes != 0 && info.name.startsWith(packageName) &&
-                        info.name != FeatureOverviewActivity::class.java.name
-                    ) {
-                        val label = getString(info.labelRes)
-                        val description = resolveString(info.descriptionRes)
-                        val category = resolveMetaData(info.metaData, metaDataKey)
-                        features.add(Feature(info.name, label, description, category!!))
-                    }
+    private fun loadFeaturesTask(vararg p0: PackageInfo?): List<Feature> {
+        val features: MutableList<Feature> = ArrayList()
+        val app = p0[0]
+        val packageName = FeatureOverviewActivity::class.java.`package`!!.name
+        val metaDataKey = getString(R.string.category)
+        if (app != null) {
+            for (info in app.activities) {
+                if (info.labelRes != 0 && info.name.startsWith(packageName) &&
+                    info.name != FeatureOverviewActivity::class.java.name
+                ) {
+                    val label = getString(info.labelRes)
+                    val description = resolveString(info.descriptionRes)
+                    val category = resolveMetaData(info.metaData, metaDataKey)
+                    features.add(Feature(info.name, label, description, category!!))
                 }
             }
-            if (!features.isEmpty()) {
-                val comparator = Comparator { lhs: Feature, rhs: Feature ->
-                    var result = lhs.category.compareTo(rhs.category, ignoreCase = true)
-                    if (result == 0) {
-                        result = lhs.getLabel().compareTo(rhs.getLabel(), ignoreCase = true)
-                    }
-                    result
+        }
+        if (features.isNotEmpty()) {
+            val comparator = Comparator { lhs: Feature, rhs: Feature ->
+                var result = lhs.category.compareTo(rhs.category, ignoreCase = true)
+                if (result == 0) {
+                    result = lhs.getLabel().compareTo(rhs.getLabel(), ignoreCase = true)
                 }
-                Collections.sort(features, comparator)
+                result
             }
-            return features
+            Collections.sort(features, comparator)
         }
+        return features
+    }
 
-        private fun resolveMetaData(bundle: Bundle?, key: String): String? {
-            var category: String? = null
-            if (bundle != null) {
-                category = bundle.getString(key)
-            }
-            return category
+    private fun resolveMetaData(bundle: Bundle?, key: String): String? {
+        var category: String? = null
+        if (bundle != null) {
+            category = bundle.getString(key)
         }
+        return category
+    }
 
-        private fun resolveString(@StringRes stringRes: Int): String {
-            return try {
-                getString(stringRes)
-            } catch (exception: NotFoundException) {
-                "-"
-            }
-        }
-
-        @RequiresApi(Build.VERSION_CODES.TIRAMISU)
-        override fun onPostExecute(features: List<Feature>) {
-            super.onPostExecute(features)
-            onFeaturesLoaded(features)
+    private fun resolveString(@StringRes stringRes: Int): String {
+        return try {
+            getString(stringRes)
+        } catch (exception: NotFoundException) {
+            "-"
         }
     }
+
     companion object {
         private const val KEY_STATE_FEATURES = "featureList"
     }