+++ platform/ios/app-swift/Sources/Styles.swift
@@ -2,6 +2,9 @@ import Foundation
 
 // #-example-code(ExampleStyles)
 
+let AMERICANA_STYLE = URL(string:
+    "https://zelonewolf.github.io/openstreetmap-americana/style.json")
+
 let VERSATILES_COLORFUL_STYLE = URL(string: "https://tiles.versatiles.org/assets/styles/colorful.json")
 
 // #-end-example-code